+++ CHANGELOG.md
@@ -17,12 +17,14 @@
 * **Application**: The new function [`Application::copyToClipboard`](https://www.boden.io/reference/foundation/application#os-services) copies a string to the global clipboard. Thanks [@mrexodia](https://github.com/mrexodia)!
 * **foundation/path**: Added the `bdn::path` namespace featuring functions to retrieve platform-specific paths like e.g. the path of the temporary directory readable/writable by the application.
 * **foundation/Application**: Added `Application` base class to the foundation module to make `Context` accessible from within the module.
+* **ui/ListView**: Added `std::optional<size_t> ListView::rowIndexForView(const std::shared_ptr<View>& view) const`
 
 ### ⚠️ Changed
 
 * **ui/View**: The View hierarchy functions have been reworked extensively. [`View::childViews`](https://www.boden.io/reference/ui/view/#view-hierarchy) returns a `std::vector` instead of a `std::list` and the parent [`View`](https://www.boden.io/reference/ui/view/) is accessible via a read-only property [`View::parentView`](https://www.boden.io/reference/ui/view/#properties). Functions that change the hierarchy have been moved out of [`View`](https://www.boden.io/reference/ui/view/).
 * **ui/View**: The `offerLayout` function has been renamed to [`setFallbackLayout`](https://www.boden.io/reference/ui/view/#layout)
-* **foudnation/Context**: Renamed `UIContext` to `Context` and moved it to foundation.
+* **foundation/Context**: Renamed `UIContext` to `Context` and moved it to foundation.
+* **ui/ListView**: Added `listView` parameter to `ListViewDataSource::viewForRowIndex`
 
 ### 🔥 Deprecated